+++ slack/src/main/java/com/business/slack/presentation/controller/SlackController.java
@@ -1,4 +1,4 @@
-package com.business.user.presentation;
+package com.business.slack.presentation.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/users")
-public class UserController {
+@RequestMapping("api/v1/slack")
+public class SlackController {
     @GetMapping("/health-check")
-    public ResponseEntity<String> healthCheck() {
+    ResponseEntity<String> healthCheck() {
         return ResponseEntity.ok("Health Check OK");
     }
 }
