+++ user/src/main/java/com/business/user/presentation/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.business.slack.presentation;
+package com.business.user.presentation.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("api/v1/slack")
-public class SlackController {
+@RequestMapping("/api/v1/users")
+public class UserController {
     @GetMapping("/health-check")
-    ResponseEntity<String> healthCheck() {
+    public ResponseEntity<String> healthCheck() {
         return ResponseEntity.ok("Health Check OK");
     }
 }
